HACK template/typekit: Fix output paths for crazy NFS semantics
The relative paths originally generated were essentially something like
BUILD_DIR/../../../SOURCE_DIR, which may not work when part of the "../"
chain includes links into an NFS file system (which is where the SOURCE_DIR
is located). This fixes the path generation to explicitly use build/source
dir paths, and no relative madness.

This is _not_ guaranteed to work for anyone else!